+++ lib/src/console_logger.dart
@@ -1,7 +1,13 @@
 import 'package:simple_dart_logger/simple_dart_logger.dart';
-import 'package:simple_dart_logger/src/log_lvl.dart';
 
-import 'logger_base.dart';
+/// Records of this type match [TermColor] to logLvls.
+typedef ConsoleLogColors = ({
+  TermColor trace,
+  TermColor debug,
+  TermColor info,
+  TermColor warning,
+  TermColor error,
+});
 
 /// This [Logger] implementation writes the log messages with a simple print
 /// statement. You should take care not to log anything in production with it
@@ -13,6 +19,14 @@ final class ConsoleLogger extends Logger {
   /// Warnings will be yellow.
   /// Errors will be red.
   final bool colorize;
+  static const ConsoleLogColors defaultColors = (
+    trace: TermColor.reset,
+    debug: TermColor.reset,
+    info: TermColor.reset,
+    warning: TermColor.yellow,
+    error: TermColor.red,
+  );
+  final ConsoleLogColors colors;
 
   /// This constructor creates a [ConsoleLogger] with a concrete configuration.
   /// If used inside a [MultiLogger] the [logLvl] and [className] properties
@@ -21,55 +35,49 @@ final class ConsoleLogger extends Logger {
   ConsoleLogger({
     super.logLvl,
     this.colorize = false,
+    this.colors = defaultColors,
     required super.className,
   });
 
   /// This constructor lacks the very important [logLvl] and [className]
   /// properties. __Only use this constructor in combination with a
   /// [MultiLogger]!__
-  factory ConsoleLogger.multi([bool colorize = false]) => ConsoleLogger(
+  factory ConsoleLogger.multi({
+    bool colorize = false,
+    ConsoleLogColors colors = defaultColors,
+  }) =>
+      ConsoleLogger(
         className: "",
         colorize: colorize,
+        colors: colors,
       );
 
   @override
-  void error(String msg) {
+  void processLogMessage(String className, String logLvl, String msg) {
     if (!colorize) {
-      super.error(msg);
+      super.processLogMessage(className, logLvl, msg);
       return;
     }
-    if (checkLogLvl(LogLvl.warning)) {
-      log(
-        TermColor.red.colorize(
-          Logger.formatter(
-            DateTime.now(),
-            "ERROR",
-            className,
-            msg,
-          ),
-        ),
-      );
-    }
-  }
 
-  @override
-  void warning(String msg) {
-    if (!colorize) {
-      super.warning(msg);
-      return;
-    }
-    if (checkLogLvl(LogLvl.warning)) {
-      log(
-        TermColor.yellow.colorize(
-          Logger.formatter(
-            DateTime.now(),
-            "WARNING",
-            className,
-            msg,
-          ),
+    final color = switch (logLvl) {
+      "TRACE" => colors.trace,
+      "DEBUG" => colors.debug,
+      "INFO" => colors.info,
+      "WARNING" => colors.warning,
+      "ERROR" => colors.error,
+      _ => TermColor.reset,
+    };
+
+    log(
+      color.colorize(
+        Logger.formatter(
+          DateTime.now(),
+          logLvl,
+          className,
+          msg,
         ),
-      );
-    }
+      ),
+    );
   }
 
   @override