+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## 1.2.2
+## 1.3.0
 
+- BUGFIX fixed issue with loglvl check when colorized logging in ConsoleLogger
+- BUGFIX colorized logging works with multilogger now
+- colors are now configurable
 - improved documentation
 - reversed version order in CHANGELOG.md